18/08/24 05:11

+++ src/main/java/com/hotel/demo/service/DetalleReservaService.java
@@ -5,33 +5,33 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.hotel.demo.interfaces.IDetalleReserva;
-import com.hotel.demo.interfacesService.IdetalleReservaService;
-import com.hotel.demo.interfacesService.IhabitacionService;
+
+
 
 import com.hotel.demo.modelo.Detalle_Reserva;
 import com.hotel.demo.modelo.Habitacion;
-
+import com.hotel.demo.modelo.Reserva;
 
 import jakarta.transaction.Transactional;
 
 @Service
-public class DetalleReservaService implements IdetalleReservaService {
+public class DetalleReservaService {
 	@Autowired
 	private IDetalleReserva data;
 	@Autowired
-	private IhabitacionService ServiceH;
+	private HabitacionService ServiceH;
 	
-	@Override
+	
 	public List<Detalle_Reserva> listarDetReserva() {
 		return (List<Detalle_Reserva>)data.findAll();
 	}
 
-	@Override
-	public Optional<Detalle_Reserva> listarId(int Id_detreserva) {
-		return data.findById(Id_detreserva);
+	
+	public Detalle_Reserva listarId(int Id_detreserva) {
+		return data.findById(Id_detreserva).orElse(new Detalle_Reserva());
 	}
 
-	@Override
+	
 	public int Guardar(Detalle_Reserva dr) {
 		int res=0;
 		Detalle_Reserva detallereserva=data.save(dr);
@@ -41,11 +41,28 @@ public class DetalleReservaService implements IdetalleReservaService {
 		return res;
 	}
 
-	@Override
-	public void Borrar(int Id_detreserva) {
-		data.deleteById(Id_detreserva);
+	
+	public Detalle_Reserva BorrarYDisponibilizar(int Id_detreserva) {
+		 Detalle_Reserva detreserva =data.findById(Id_detreserva).orElse(null);
+	        if (detreserva !=null) {
+	            // Obtener la reserva asociada al detalle
+	            Reserva reserva = detreserva.getReserva();
+
+	            // Verificar si la reserva y la habitación asociada existen
+	            if (reserva != null && reserva.getHabitacion() != null) {
+	                // Obtener la habitación y actualizar su estado
+	                Habitacion habitacion = reserva.getHabitacion();
+	                habitacion.disponibilizar(); // Utiliza el método que actualiza el estado
+	                ServiceH.Guardar(habitacion); // Guardar la habitación
+	            }
+
+	            // Eliminar el detalle de la reserva
+	            data.deleteById(Id_detreserva);
+	            return detreserva;
+	        }
+			return detreserva;
 	}
-	@Override
+	
 	public Detalle_Reserva GuardarDR(Detalle_Reserva dr) {
 		    // Establecer un valor predeterminado si id_emp es null
 		    
@@ -56,7 +73,7 @@ public class DetalleReservaService implements IdetalleReservaService {
 	@Transactional
 	public Detalle_Reserva crearReservaConDetalle(Detalle_Reserva detreserva) {
 	    // Obtener la habitación por ID
-	    Habitacion habitacion = ServiceH.listarNro(detreserva.getNro_habi()).orElse(null);
+	    Habitacion habitacion = ServiceH.listarNro(detreserva.getHabitacion().getNro_habi());
 
 	    if (habitacion != null) {
 	        // Guardar la reserva
@@ -65,7 +82,7 @@ public class DetalleReservaService implements IdetalleReservaService {
 	        // Configurar el objeto Detalle_Reserva
 	        
 	        
-	        detreserva.setObjHabitacion(habitacion);
+	        detreserva.setHabitacion(habitacion);
 	        // Configurar otros campos según sea necesario
 
 	        // Guardar el detalle de la reserva
